[K/N][IR] Removed KonanBackendContext
`Context` is used everywhere instead. It will be renamed later to
`NativeBackendContext`.

^KT-72903
