--- kotlin-native/backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/KonanBackendContext.kt
@@ -1,44 +1,0 @@
-/*
- * Copyright 2010-2024 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package org.jetbrains.kotlin.backend.konan
-
-import org.jetbrains.kotlin.K1Deprecation
-import org.jetbrains.kotlin.backend.common.CommonBackendContext
-import org.jetbrains.kotlin.backend.common.InlineClassesUtils
-import org.jetbrains.kotlin.backend.common.ir.KlibSharedVariablesManager
-import org.jetbrains.kotlin.backend.konan.driver.BasicNativeBackendPhaseContext
-import org.jetbrains.kotlin.backend.konan.ir.BackendNativeSymbols
-import org.jetbrains.kotlin.builtins.konan.KonanBuiltIns
-import org.jetbrains.kotlin.descriptors.ValueClassBackendAgnosticApi
-import org.jetbrains.kotlin.ir.declarations.IrClass
-import org.jetbrains.kotlin.ir.declarations.IrFactory
-import org.jetbrains.kotlin.ir.declarations.impl.IrFactoryImpl
-import org.jetbrains.kotlin.ir.declarations.isInlineClass
-
-internal abstract class KonanBackendContext(config: NativeSecondStageCompilationConfig) : BasicNativeBackendPhaseContext(config), CommonBackendContext {
-    @OptIn(ValueClassBackendAgnosticApi::class)
-    override val inlineClassesUtils: InlineClassesUtils = object : InlineClassesUtils {
-        override fun isClassInlineLike(klass: IrClass): Boolean =
-                klass.isInlineClass(treatCompatibleFullValueClassesAsInline = true)
-    }
-
-    @OptIn(K1Deprecation::class)
-    abstract val builtIns: KonanBuiltIns
-
-    abstract override val symbols: BackendNativeSymbols
-
-    override val sharedVariablesManager by lazy {
-        // Creating lazily because builtIns module seems to be incomplete during `link` test;
-        // TODO: investigate this.
-        KlibSharedVariablesManager(symbols)
-    }
-
-    override val irFactory: IrFactory = IrFactoryImpl
-
-    override fun log(message: String) {
-        super<BasicNativeBackendPhaseContext>.log(message)
-    }
-}
